requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
psycopg==3.2.4
psycopg-binary==3.2.4
sqlparse==0.5.3
